import and delete many

+++ PRN231-StudentManagmentSystem/StudentManagingSystem-Client/Pages/DepartmentPage/DeleteDepartment.cshtml.cs
@@ -3,17 +3,33 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using StudentManagingSystem_Client.Services;
+using StudentManagingSystem_Client.ViewModel;
 
 namespace StudentManagingSystem_Client.Pages.DepartmentPage
 {
     [Authorize(Roles = RoleConstant.ADMIN)]
     public class DeleteDepartmentModel : PageModel
     {
-        public async Task<IActionResult> OnGetAsync(Guid id)
+        [BindProperty]
+        public List<string> SelectedItems { get; set; }
+        public async Task<IActionResult> OnGetAsync(Guid id, List<string> selectedItems)
         {
             var client = new ClientService(HttpContext);
-            var res = await client.Delete("/api/Department/delete", $"?id={id}");
-            return RedirectToPage("/DepartmentPage/Department");
+            if (SelectedItems != null && SelectedItems.Count > 0)
+            {
+                var requestModel = new DepartmentDeleteListRequest
+                {
+                    Id = SelectedItems
+                };
+                var res1 = await client.PostAdd("/api/Department/deleteList", requestModel);
+                return RedirectToPage("/DepartmentPage/Department");
+            }
+            else
+            {
+                var res = await client.Delete("/api/Department/delete", $"?id={id}");
+                return RedirectToPage("/DepartmentPage/Department");
+            }
+            
         }
     }
 }
